services/web/requirements.txt
# Core Dependencies
Django==3.2.7
djangorestframework==3.12.4
elasticsearch==7.15.0
gunicorn==20.1.0
psycopg2-binary==2.9.1
django-elasticsearch-dsl==7.2.1
django-filter==21.1
django-registration==3.2
django-bootstrap4==21.1
django-hCaptcha==0.2.0
PyJWT==2.3.0
django-select2==7.9.0
django-addanother==2.0.2
django-extra-views==0.14.0
django-easy-audit==1.3.2
pinax-messages==3.0.0
django-invitations==1.9.3
pinax-announcements==4.0.0
stream-framework[redis]==1.4.0
celery[redis]==5.2.3
django-celery-beat==2.2.1
django-celery-results==2.3.0
gevent==21.12.0
rules==3.3

# Geographic analysis and visualisation
django-leaflet==0.28.2

# Django REST Framework dependencies for OpenAPI schema generation
PyYAML==6.0
uritemplate==4.1.1

# Crawling websites and APIs
Scrapy==2.5.1
deepdiff==5.6.0

# Admin formatting
Pygments==2.10.0

# Calling external API services
requests==2.26.0
django-anymail==8.4

# Text analysis
textacy==0.11.0

# Dev dependencies
ipython==7.29.0
django-debug-toolbar==3.2.2
django-extensions==3.1.5
Werkzeug==2.0.2
black==21.11b1
selenium==4.1.0
ipdb==0.13.9
tblib==1.7.0
